feat:add function to next page
