+++ mainapp/modules/security_utils.py
@@ -35,3 +35,5 @@ def check_password():
     else:
         # Password is correct
         return True
+
+
